Added auth and multi-tenancy(routers)

+++ studies_api/routers/study_sessions.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, status, Depends, HTTPException, Query
 from typing import Optional
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import select
+from sqlalchemy.orm import selectinload
 
 from studies_api.schemas.study_sessions import (
     StudySessionListPublicSchema,
@@ -10,7 +11,10 @@ from studies_api.schemas.study_sessions import (
     StudySessionUpdateSchema
 )
 from studies_api.models.sessions import Session
+from studies_api.models.users import User
 from studies_api.core.database import get_connection
+from studies_api.core.security import get_current_user, verify_study_session_ownership
+
 
 router = APIRouter()
 
@@ -21,7 +25,11 @@ router = APIRouter()
     response_model=StudySessionPublicSchema,
     summary='Create new Study Session',
 )
-async def create_session(session: StudySessionSchema, db: AsyncSession = Depends(get_connection)):
+async def create_session(
+    session: StudySessionSchema,
+    db: AsyncSession = Depends(get_connection),
+    current_user: User = Depends(get_current_user),
+    ):
     db_session = Session(
         topic=session.topic,
         duration_minutes=session.duration_minutes,
@@ -47,6 +55,7 @@ async def list_sessions(
     offset: int = Query(0, ge=0, description='Number of registers to skip'),
     limit: int = Query(100, ge=1, le=100, description='Limit of registers'),
     search: Optional[str] = Query(None, description='Search by Topic'),
+    current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_connection)):
     query = select(Session)
 
@@ -72,43 +81,75 @@ async def list_sessions(
     path="/study-session/{session_id}", 
     status_code=status.HTTP_200_OK,
     response_model=StudySessionPublicSchema,
-    summary="Search Study Session by ID"
+    summary="Search Study Session by ID",
     )
-async def get_session(session_id: int, db: AsyncSession = Depends(get_connection)):
-    study_session = await db.get(Session, session_id)
+async def get_session(
+    session_id: int,
+    db: AsyncSession = Depends(get_connection),
+    current_user: User = Depends(get_current_user),
+    ):
+    query = select(Session).where(Session.id == session_id)
+    result = await db.execute(query)
+
+    study_session = result.scalar_one_or_none()
 
     if not study_session:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Study Session Not Found",
             )
+    
+    verify_study_session_ownership(study_session=study_session, current_user=current_user)
 
     return study_session
 
 
 @router.put(
     path='/study-session/{session_id}',
-    status_code=status.HTTP_201_CREATED,
+    status_code=status.HTTP_200_OK,
     response_model=StudySessionPublicSchema,
     summary='Update Study Session'
 )
-async def update_session(session_id: int, session_update: StudySessionUpdateSchema, db: AsyncSession = Depends(get_connection)):
-    session = await db.get(Session, session_id)
+async def update_session(
+    session_id: int,
+    session_update: StudySessionUpdateSchema, 
+    db: AsyncSession = Depends(get_connection),
+    current_user: User = Depends(get_current_user),
+    ):
+    study_session = await db.get(Session, session_id)
 
-    if not session:
+    if not study_session:
         raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
+            status_code=status.HTTP_404_NOT_FOUND,
             detail='Study Session Not Found'
         )
     
+    verify_study_session_ownership(study_session=study_session, current_user=current_user)
+    
+    update_data = session_update.model_dump(exclude_unset=True)
+
+    for field, value in update_data.items():
+        setattr(study_session, field, value)
+    
+    await db.commit()
+    await db.refresh(study_session)
+    
+    return study_session
+    
     
 
-
 @router.delete(path='/study-session/{session_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_session(session_id: int, db: AsyncSession = Depends(get_connection)):
+async def delete_session(
+    session_id: int, 
+    db: AsyncSession = Depends(get_connection),
+    current_user: User = Depends(get_current_user),
+    ):
     study_session = await db.get(Session, session_id)
+
     if not study_session:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Study Session Not Found')
+    
+    verify_study_session_ownership(study_session=study_session, current_user=current_user)
 
     await db.delete(study_session)
     await db.commit()
